refactor(academy): move certificate-validity into application/util (AYC-597)
Pure application-layer helpers live in an application/util/ folder in this
repo (iam/subscriptions/application/util/ is the closest analogue — same
kind of date-window math), not loose at the application root.
