+++ ayunis-core-backend/src/domain/academy/application/util/certificate-validity.ts
